src/affiliate_marketing.py
```python
from pymongo import MongoClient
from bson.objectid import ObjectId

# MongoDB connection
client = MongoClient('mongodb://localhost:27017/')
db = client['ai_shopper_db']

# Collection for affiliate partners
affiliatePartners = db['affiliate_partners']

# Schema for affiliate marketing data
AffiliateSchema = {
    'partner_name': str,
    'commission_rate': float,
    'affiliate_link': str
}

def initiateAffiliateMarketing(userProfileData):
    """
    Function to initiate affiliate marketing.
    It selects the best affiliate partner based on the user's preferences and the commission rate.
    """
    best_partner = None
    max_commission = 0

    for partner in affiliatePartners.find():
        if partner['partner_name'] in userProfileData['preferences']:
            if partner['commission_rate'] > max_commission:
                max_commission = partner['commission_rate']
                best_partner = partner

    if best_partner:
        return best_partner['affiliate_link']
    else:
        return None

def addAffiliatePartner(partner_name, commission_rate, affiliate_link):
    """
    Function to add a new affiliate partner.
    """
    new_partner = {
        'partner_name': partner_name,
        'commission_rate': commission_rate,
        'affiliate_link': affiliate_link
    }

    # Validate data against schema
    for key in new_partner.keys():
        if type(new_partner[key]) != AffiliateSchema[key]:
            raise ValueError(f"Incorrect type for {key}. Expected {AffiliateSchema[key]} but got {type(new_partner[key])}")

    # Insert new partner into the database
    affiliatePartners.insert_one(new_partner)

def removeAffiliatePartner(partner_id):
    """
    Function to remove an affiliate partner.
    """
    affiliatePartners.delete_one({'_id': ObjectId(partner_id)})
```